polls: move from assets/ to static/ for consistency
