+++ adhocracy4/polls/static/PollDetail/PollOpenQuestion.jsx
