+++ adhocracy4/polls/static/PollDetail/PollQuestion.jsx
