+++ adhocracy4/polls/static/__tests__/__testdata__/QUESTION_OBJECT.js
