+++ adhocracy4/polls/static/react_polls.jsx
